Remove unnecessaries
