+++ ftc/controllers/Flat/flat.py
@@ -1,25 +1,36 @@
 """ References
 [1] M. Faessler, A. Franchi and D. Scaramuzza, "Differential Flatness of Quadrotor Dynamics Subject to Rotor Drag for Accurate Tracking of High-Speed Trajectories," in IEEE Robotics and Automation Letters, vol. 3, no. 2, pp. 620-626, April 2018, doi: 10.1109/LRA.2017.2776353.
 """
-
-import fym
+import matplotlib.pyplot as plt
 import numpy as np
-from fym.utils.rot import quat2angle, quat2dcm
-from numpy import cos, sin, tan
+from numpy import cos, sin
 
 
-class FlatController(fym.BaseEnv):
-    def __init__(self, env):
+class FlatController:
+    def __init__(self, m, g, J):
         super().__init__()
-        self.m = env.plant.m
-        self.g = env.plant.g * np.vstack((0, 0, 1))
-        self.J = env.plant.J
+        self.m = m
+        self.g = g * np.vstack((0, 0, 1))
+        self.J = J
 
-    def get_control(self, t, env):
-        # pos, vel, quat, omega = env.plant.observe_list()
-        # ang = np.vstack(quat2angle(quat)[::-1])
+    def get_ref(self, t, *args):
+        # desired trajectories
+        posd = np.vstack([np.sin(t), np.cos(t), -t])
+        posd_1dot = np.vstack([np.cos(t), -np.sin(t), -1])
+        posd_2dot = np.vstack([-np.sin(t), -np.cos(t), 0])
+        posd_3dot = np.vstack([-np.cos(t), np.sin(t), 0])
+        posd_4dot = np.stack([np.sin(t), np.cos(t), 0])
+        refs = {
+            "posd": posd,
+            "posd_1dot": posd_1dot,
+            "posd_2dot": posd_2dot,
+            "posd_3dot": posd_3dot,
+            "posd_4dot": posd_4dot,
+        }
+        return [refs[key] for key in args]
 
-        posd, veld, accd, jerkd, snapd = env.get_ref(
+    def get_control(self, t):
+        posd, veld, accd, jerkd, snapd = self.get_ref(
             t, "posd", "posd_1dot", "posd_2dot", "posd_3dot", "posd_4dot"
         )
         psid = 0
@@ -42,7 +53,7 @@ class FlatController(fym.BaseEnv):
         cdot = -np.dot(Zb.T, jerkd)
         Fz = self.m * c
 
-        # angular rates
+        # Angular rates
         p = np.dot(Yb.T, jerkd) / c
         q = -np.dot(Xb.T, jerkd) / c
         r = (psid_1dot * Xc.T @ Xb + q * Yc.T @ Zb) / np.linalg.norm(
@@ -50,7 +61,7 @@ class FlatController(fym.BaseEnv):
         )
         omega = np.vstack((p, q, r))
 
-        # derivatives of angular rates
+        # Derivatives of angular rates
         pdot = Yb.T @ snapd / c - 2 * cdot * (Yb.T @ jerkd / c**2) + q * r
         qdot = -Xb.T @ snapd / c + 2 * cdot * (Xb.T @ jerkd / c**2) - p * r
         rdot = (
@@ -64,11 +75,53 @@ class FlatController(fym.BaseEnv):
 
         M = self.J @ omega_dot + np.cross(omega.T, (self.J @ omega).T).T
 
-        FM_traj = np.vstack((0, 0, Fz, M))
-        controller_info = {
-            "posd": posd,
-            "veld": veld,
-            "psid": psid,
-            "omegad": omega,
-        }
-        return FM_traj, controller_info
+        FM = np.vstack((0, 0, Fz, M))
+        return FM
+
+
+if __name__ == "__main__":
+    m, g = 1, 9.81
+    J = np.diag([1, 1, 1])
+
+    tspan = np.linspace(0, 20, 200)
+    ctrl = FlatController(m, g, J)
+    FM_traj = np.empty((6, 0))
+    for t in tspan:
+        FM = ctrl.get_control(t)
+        FM_traj = np.append(FM_traj, FM, axis=1)
+
+    """Figure - Forces & Moments trajectories"""
+    fig, axes = plt.subplots(3, 2, squeeze=False, sharex=True)
+
+    """ Column 1 - Forces """
+    ax = axes[0, 0]
+    ax.plot(tspan, FM_traj[0, :], "k-")
+    ax.set_ylabel(r"$F_x$")
+    ax.legend(["Response"], loc="upper right")
+
+    ax = axes[1, 0]
+    ax.plot(tspan, FM_traj[1, :], "k-")
+    ax.set_ylabel(r"$F_y$")
+
+    ax = axes[2, 0]
+    ax.plot(tspan, FM_traj[2, :], "k-")
+    ax.set_ylabel(r"$F_z$")
+    ax.set_xlabel("Time, sec")
+
+    """ Column 2 - Moments """
+    ax = axes[0, 1]
+    ax.plot(tspan, FM_traj[3, :], "k-")
+    ax.set_ylabel(r"$M_x$")
+
+    ax = axes[1, 1]
+    ax.plot(tspan, FM_traj[4, :], "k-")
+    ax.set_ylabel(r"$M_y$")
+
+    ax = axes[2, 1]
+    ax.plot(tspan, FM_traj[5, :], "k-")
+    ax.set_ylabel(r"$M_z$")
+    ax.set_xlabel("Time, sec")
+
+    plt.tight_layout()
+
+    plt.show()
